+++ miamCore/src/commonMain/kotlin/com/miam/kmmMiamCore/component/pricing/PricingContract.kt
@@ -9,22 +9,13 @@ import com.miam.kmmMiamCore.miam_core.model.Pricing
 
 interface PricingContract {
 
-    sealed class Event : UiEvent {
-        data class OnSetRecipe(val idRecipe: String, val guestNumber: Int) : Event()
-        data class SetPrice(val pricing: Pricing) : Event()
-        data class SetDirectPrice(val price: Double) : Event()
-        object OnPriceUpdate : Event()
-    }
+    sealed class Event: UiEvent
 
     data class State(
         val price: BasicUiState<Pricing>,
-        val directPrice: Double?,
-        val isInCart: Boolean,
-        val recipeId: String,
-        val guestNumber: Int,
-        val integerPart: String,
-        val decimalPart: String
-    ) : UiState
+        val recipeId: String? = null,
+        val guestNumber: Int? = null
+    ): UiState
 
-    sealed class Effect : UiEffect
+    sealed class Effect: UiEffect
 }