+++ changelog.md
@@ -1,5 +1,6 @@
 # 3.3.0
 - [FEA][COR] Clean counter usage
+- [FEA][COR] Optimize price component
 - [FIX][AND] Add missing templates in like button and preferences search
 - [FIX][IOS] Unwrapp AsyncImage URL and replace placeholder 
 - [FIX][IOS] Add function preferenceTapped to catalogViewToolbarTemplate